+++ src/main/java/com/countmyh/model/VacationEntry.java
@@ -1,3 +1,3 @@
 package com.countmyh.model;
 
-public record VacationEntry(int year, int month, int days) {}
+public record VacationEntry(int year, int month, double days) {}
